+++ src/utils/context.ts
@@ -1,4 +1,5 @@
 import * as vscode from 'vscode';
+import * as logger from './logger';
 
 let _context: vscode.ExtensionContext | undefined;
 
@@ -8,7 +9,9 @@ export function set(context: vscode.ExtensionContext) {
 
 export function get(): vscode.ExtensionContext {
     if (!_context) {
-        throw new Error("Context not initialized! Make sure to call set_context in activate().");
+        const err = new Error("Context not initialized! Make sure to call set_context in activate().");
+        logger.error(err);
+        throw err;
     }
     return _context;
 }